Introduced a new modular save system. Refactored project context installers by splitting the large installer into smaller classes. Fixed a bug with unloading the previous scene caused by a race condition. Renamed components for better clarity.

+++ unity-game-template-project/Assets/Modules/Wallets/Scripts/Types/CurrencyValue.cs
@@ -5,7 +5,7 @@ namespace Modules.Wallet.Types
     [System.Serializable]
     public class CurrencyValue
     {
-        public CurrencyValue(CurrencyType currencyType, long amount) 
+        public CurrencyValue(CurrencyType currencyType, int amount) 
         {
             CurrencyType = currencyType;
             Amount = amount;
@@ -13,6 +13,6 @@ namespace Modules.Wallet.Types
 
         [field: SerializeField] public CurrencyType CurrencyType { get; private set; }
 
-        [field: SerializeField] public long Amount { get; private set; }
+        [field: SerializeField] public int Amount { get; private set; }
     }
 }
